helmet-detection-streaming-app/requirements.txt
streamlit
opencv-python
numpy
Pillow
ultralytics
flask
flask-socketio
eventlet